requirements.txt
requests>=2.28.0
openpyxl>=3.1.0
Pillow>=10.0.0
